refactor: security hardening and DRY improvements for DocuSign connector
- Fix silent account_id retrieval: all 7 actions now use requireAccountID()
  which returns a ValidationError instead of silently using an empty string
- Add base_url SSRF prevention: validateBaseURL() ensures the custom URL
  points to a *.docusign.net host over HTTPS only
- Extract parseParams() helper with validatable interface to eliminate
  repeated unmarshal/validate/getAccountID boilerplate across all actions
- Split docusign.go (587 lines) into focused files:
  docusign.go (core), docusign_http.go, docusign_errors.go, docusign_manifest.go
- Fix race condition in TestCheckStatus_WithoutRecipients using sync.Mutex
  and atomic counter for thread-safe request tracking
- Add comprehensive tests for base_url validation and missing account_id
