+++ connectors/docusign/check_status.go
@@ -2,8 +2,6 @@ package docusign
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
 	"net/url"
 
 	"github.com/supersuit-tech/permission-slip-web/connectors"
@@ -69,14 +67,10 @@ type recipientDetail struct {
 
 func (a *checkStatusAction) Execute(ctx context.Context, req connectors.ActionRequest) (*connectors.ActionResult, error) {
 	var params checkStatusParams
-	if err := json.Unmarshal(req.Parameters, &params); err != nil {
-		return nil, &connectors.ValidationError{Message: fmt.Sprintf("invalid parameters: %v", err)}
-	}
-	if err := params.validate(); err != nil {
+	accountID, err := parseParams(req, &params)
+	if err != nil {
 		return nil, err
 	}
-
-	accountID, _ := req.Credentials.Get(credKeyAccountID)
 
 	var resp checkStatusResponse
 	path := accountPath(accountID) + "/envelopes/" + url.PathEscape(params.EnvelopeID)
